requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.6.0
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
#mod-wsgi==4.7.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyPDF2==1.26.0
pytz==2020.1
sqlparse==0.4.1
